fix: move `Type` class to `empaphy\usephul\var` namespace
BREAKING CHANGE: this changes the namespace of `Type`
